newsletter_agency/ResearchAgent/tools/TrendAnalysisTool.py
from agency_swarm.tools import BaseTool
from pydantic import Field
import os
import requests
from dotenv import load_dotenv

load_dotenv()

class TrendAnalysisTool(BaseTool):
    """
    Analyzes cutting-edge technologies and methodologies in athlete development,
    talent identification, and coaching innovations. Focuses on tools and strategies
    that enhance performance, recovery, and scouting effectiveness.
    """
    
    topic: str = Field(
        ..., 
        description="Specific area in athlete development (e.g., 'Biomechanics tech', 'Recovery innovations', 'AI scouting tools')"
    )
    
    def run(self):
        # Add novelty scoring
        analysis = {
            "novelty_score": self._calculate_novelty_score(),
            "industry_impact": self._assess_industry_impact(),
            "athlete_benefits": self._identify_athlete_advantages(),
            "counterintuitive_insights": self._find_unexpected_angles()
        }
        return f"""
        **Emerging Trend:** {self.topic}
        - Novelty Potential: {analysis['novelty_score']}/10
        - Key Insight: {analysis['counterintuitive_insights']}
        - Athlete Impact: {analysis['athlete_benefits']}
        - Strategic Value: {analysis['industry_impact']}
        """

    def _calculate_novelty_score(self):
        # Implementation of _calculate_novelty_score method
        pass

    def _assess_industry_impact(self):
        # Implementation of _assess_industry_impact method
        pass

    def _identify_athlete_advantages(self):
        # Implementation of _identify_athlete_advantages method
        pass

    def _find_unexpected_angles(self):
        # Implementation of _find_unexpected_angles method
        pass

if __name__ == "__main__":
    tool = TrendAnalysisTool(topic="NBA player development tech")
    print(tool.run()) 