newsletter_agency/requirements.txt
requests
python-dotenv
anthropic>=0.7.0 